+++ x-pack/solutions/security/plugins/security_solution/public/flyout/document_details/left/components/graph_visualization.tsx
@@ -9,9 +9,9 @@ import React, { memo, useMemo } from 'react';
 import { buildDataTableRecord, type EsHitRecord } from '@kbn/discover-utils';
 import { useDocumentDetailsContext } from '../../shared/context';
 import { useGraphPreview } from '../../../../flyout_v2/document/hooks/use_graph_preview';
-import { GraphVisualization as SharedGraphVisualization } from '../../../shared/components/graph_visualization';
+import { GraphVisualization as SharedGraphVisualization } from '../../../../flyout_v2/graph/components/graph_visualization';
 
-export { GRAPH_ID } from '../../../shared/components/graph_visualization';
+export { GRAPH_ID } from '../../../../flyout_v2/graph/components/graph_visualization';
 
 /**
  * Graph visualization view displayed in the document details expandable flyout left section under the Visualize tab.
